requirements.txt
# 基础依赖
# tkinter是Python内置库，无需在requirements中声明
PyYAML >= 6.0.1         # 配置加载需要
# deepseek-sdk >= 0.1.0   # 当前可用版本

# 开发依赖（可选）
black == 23.3.0         # 代码格式化
pylint == 2.17.4        # 代码质量检查 